package improvements
