+++ echo-typescript-sdk/src/index.ts
@@ -1,4 +1,5 @@
 export { EchoClient } from './client';
-export { EchoConfig } from './config';
+export * from './auth';
+export * from './cli-helpers';
 export * from './types';
 export { EchoDbService, DatabaseClient } from './db-service';
